Rewrite reducers with slices
