+++ vanguarde/src/reducers/songReducer.js
@@ -1,8 +1,8 @@
-import {createReducer, current} from '@reduxjs/toolkit';
+import {current, createSlice} from '@reduxjs/toolkit';
 
-import {fetchBegin, fetchError} from '../actions/fetching';
+import axios from 'axios';
 
-import {fetchSongListSuccess} from '../actions/songList';
+import {handleError, handleSuccess} from '../actions/fetching';
 
 const initialState = {
     songList: [],
@@ -10,24 +10,48 @@ const initialState = {
     error: null,
 };
 
-const songReducer = createReducer(
+const songSlice = createSlice({
+    name: 'songs',
     initialState,
-    {
-        [fetchBegin]: (state, action) => {
+    reducers: {
+        fetchBegin: (state, action) => {
             state.loading = true;
             state.error = null;
         },
-        [fetchSongListSuccess]: (state, action) => {
-            const songList = action.payload.songList;
+        fetchSongListSuccess: (state, action) => {
+            const songList = action.payload;
             state.loading = false;
             state.songList = songList;
             console.log('new state: ', current(state));
         },
-        [fetchError]: (state, action) => {
+        fetchError: (state, action) => {
             state.loading = false;
             state.error = action.payload.error;
         },
     },
-);
+});
 
-export default songReducer;
+export function makeSongQuery(dispatch, getState, {apiConfig}, searchQuery = {}) {
+    dispatch(fetchBegin());
+    console.log({searchQuery});
+    axios.get(`${apiConfig.url}/songs`, {params: searchQuery})
+        .then((res) => {
+            dispatch(fetchSongListSuccess(res.data));
+            dispatch(handleSuccess(res));
+        })
+        .catch((error) => {
+            dispatch(fetchError(error));
+            dispatch(handleError(error));
+        });
+}
+
+export function getSongList(searchQuery = {}) {
+    return (dispatch, getState, {apiConfig}) => {
+        makeSongQuery(dispatch, getState, {apiConfig}, searchQuery);
+    };
+}
+
+
+const {actions, reducer} = songSlice;
+export const {fetchBegin, fetchSongListSuccess, fetchError} = actions;
+export default reducer;
